requirements.txt
google-genai>=1.0.0
langdetect>=1.0.9
trafilatura>=2.0.0
feedparser>=6.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pyairtable>=2.3.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.126.0
